Added export option to add a descriptive extension.
Because I am inconsistent in my profile management I often tweak a setting of a profile before printing.  This means the profile name usually does not reflect the actual profile values.  This commit adds an extension to the gcode file with some key profile values. If a module is turned off then the value is not added to the extension.

E.g. <filename>.04hx06w_33Ff_03fill.4off.gcode

where:

04h    = "Layer Thickness (mm):" 0.4
x      =
06w    = 0.6 width i.e. 0.4 times "Perimeter Width over Thickness (ratio):" 1.5
_
33Ff   = "Flow Rate Setting (float):" 33.0 and "Feed Rate (mm/s):" 33.0.  If either value has a positive value after the decimal place then this is also shown, but if it is zero it is hidden.  Also, if the values differ (which they shouldn't with 5D volumetrics) then each is displayed seperately, i.e. 30F35.2f = "Flow Rate Setting (float):" 30.0 and "Feed Rate (mm/s):" 35.2.
_
03fill = "Infill Solidity (ratio):" 0.3
.
4off   = "Number of Columns (integer):" x "Number of Rows (integer):" from the Multiply Tab, e.g. 2x2.  If these values are both 1 then this part is not shown.